# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /master-products/source-categories:
    get:
      operationId: getMasterProductSourceCategories
      tags: [masterProducts]
      summary: Read the upstream catalogue vocabulary without mutating local data
      responses:
        "200":
          description: Distinct upstream divisions and categories with mapping status
          content:
            application/json:
              schema: { type: object }
        "502":
          description: Upstream catalogue unavailable or invalid
...
  /master-products/sync:
    post:
      operationId: syncMasterProducts
      tags: [masterProducts]
      summary: Sync the maintained catalogue copy from Prayag-Competition-Analysis
      responses:
        "200":
          description: Sync counts and source vocabulary report
          content:
            application/json:
              schema: { type: object }
        "502": { description: Upstream catalogue unavailable or sync failed }
...
  /master-products/products:
    get:
      operationId: listMasterProducts
      tags: [masterProducts]
      summary: List the complete product review roster for a planning segment
      parameters:
        - name: segment
          in: query
          required: false
          schema: { type: string, enum: [PTMT, Plumbing, CP], default: PTMT }
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [classified, unclassified, ambiguous] }
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: source
          in: query
          required: false
          schema: { type: string, enum: [workbook, catalogue, seed] }
        - name: search
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Products with classification, source, demand, and buffer provenance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductListResponse" }
        "400": { description: Invalid filter }
        "500": { description: Product roster unavailable }
  /master-products/products/{itemCode}/{colour}:
    patch:
      operationId: reclassifyMasterProduct
      tags: [masterProducts]
      summary: Reclassify a planning-roster product with an audit reason
      parameters:
        - name: itemCode
          in: path
          required: true
          schema: { type: string }
        - name: colour
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductReclassificationInput" }
      responses:
        "200": { description: Product reclassified }
        "400": { description: Invalid reclassification }
        "404": { description: Product is not in the planning roster }
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      parameters:
        - name: month
          in: query
          required: true
          schema: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
        - name: segment
          in: query
          required: true
          schema: { type: string, enum: [PTMT, Plumbing] }
      responses:
        "200":
          description: Evaluated alert rules for one segment and month
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertsResponse" }
        "400": { description: Invalid alert scope }
        "502": { description: Alert evaluation failed }
  /alerts/history:
    get:
      operationId: getAlertsHistory
      tags: [alerts]
      parameters:
        - name: segment
          in: query
          required: true
          schema: { type: string, enum: [PTMT, Plumbing] }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Alert state-change history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertHistoryResponse" }
  /alerts/{id}/acknowledge:
    patch:
      operationId: acknowledgeAlert
      tags: [alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Alert acknowledged
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertActionResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /alerts/{id}/mute:
    patch:
      operationId: muteAlert
      tags: [alerts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MuteAlertRequest" }
      responses:
        "200":
          description: Alert muted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertActionResponse" }
        "400": { description: Invalid mute request }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /alerts/thresholds/{code}:
    put:
      operationId: updateAlertThreshold
      tags: [alerts]
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, enum: [R1, R2, R3, R4, R5, R6, R7] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAlertThresholdRequest" }
      responses:
        "200":
          description: Threshold updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertThresholdActionResponse" }
        "400": { description: Invalid threshold }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /alerts/thresholds/{code}/reset:
    post:
      operationId: resetAlertThreshold
      tags: [alerts]
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string, enum: [R1, R2, R3, R4, R5, R6, R7] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [segment]
              properties:
                segment: { type: string, enum: [PTMT, Plumbing] }
      responses:
        "200":
          description: Threshold reset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertThresholdActionResponse" }
components:
  securitySchemes:
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: "Managed API key with read scope and optional PTMT/Plumbing segment restrictions."
...
  schemas:
...
    ProductListRow:
      type: object
      required: [key, segment, itemCode, colour, productName, division, category, status, source, note, inCatalogue, inPlanningWorkbook, lastSeenProductionMonth, pendingQuantity, dummyQuantity, bufferReq, auditCount]
      properties:
        key: { type: string }
        segment: { type: string, enum: [PTMT, Plumbing, CP] }
        itemCode: { type: string }
        colour: { type: ["string", "null"] }
        productName: { type: ["string", "null"] }
        division: { type: ["string", "null"] }
        category: { type: string }
        status: { type: string, enum: [classified, unclassified, ambiguous] }
        source: { type: ["string", "null"], enum: [workbook, catalogue, seed, null] }
        note: { type: ["string", "null"] }
        inCatalogue: { type: boolean }
        inPlanningWorkbook: { type: boolean }
        lastSeenProductionMonth: { type: ["string", "null"] }
        pendingQuantity: { type: number }
        dummyQuantity: { type: number }
        bufferReq: { type: ["number", "null"] }
        auditCount: { type: integer }
    ProductListResponse:
      type: object
      required: [rows, total, categories]
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/ProductListRow" }
        total: { type: integer }
        categories:
          type: array
          items: { type: string }
    ProductReclassificationInput:
      type: object
      required: [segment, category, status, reason]
      properties:
        segment: { type: string, enum: [PTMT, Plumbing, CP] }
        category: { type: string }
        status: { type: string, enum: [classified, unclassified, ambiguous] }
        reason: { type: string }
    AlertState:
      type: string
      enum: [clear, fired, muted, suppressed]
    AlertSeverity:
      type: string
      enum: [red]
    AlertSourceLink:
      type: object
      required: [label, href]
      properties:
        label: { type: string }
        href: { type: string }
    Alert:
      type: object
      required: [id, code, segment, month, severity, state, title, action, message, value, threshold, difference, quantity, details, sourceLinks, firstSeenAt, lastEvaluatedAt, acknowledgedAt, acknowledgedBy, mutedUntil, muteReason, suppressedReason]
      properties:
        id: { type: ["integer", "null"] }
        code: { type: string, enum: [R1, R2, R3, R4, R5, R6, R7] }
        segment: { type: string, enum: [PTMT, Plumbing] }
        month: { type: string }
        severity: { $ref: "#/components/schemas/AlertSeverity" }
        state: { $ref: "#/components/schemas/AlertState" }
        title: { type: string }
        action: { type: string }
        message: { type: string }
        value: { type: ["number", "null"] }
        threshold: { type: ["number", "null"] }
        difference: { type: ["number", "null"] }
        quantity: { type: number }
        details: { type: object, additionalProperties: true }
        sourceLinks:
          type: array
          items: { $ref: "#/components/schemas/AlertSourceLink" }
        firstSeenAt: { type: ["string", "null"], format: date-time }
        lastEvaluatedAt: { type: string, format: date-time }
        acknowledgedAt: { type: ["string", "null"], format: date-time }
        acknowledgedBy: { type: ["string", "null"] }
        mutedUntil: { type: ["string", "null"], format: date-time }
        muteReason: { type: ["string", "null"] }
        suppressedReason: { type: ["string", "null"] }
    AlertThreshold:
      type: object
      required: [code, segment, name, description, value, defaultValue, unit, scope, observedMin, observedMax, wouldFireCount, updatedBy, updatedAt]
      properties:
        code: { type: string, enum: [R1, R2, R3, R4, R5, R6, R7] }
        segment: { type: string, enum: [PTMT, Plumbing] }
        name: { type: string }
        description: { type: string }
        value: { type: number }
        defaultValue: { type: number }
        unit: { type: string }
        scope: { type: string }
        observedMin: { type: ["number", "null"] }
        observedMax: { type: ["number", "null"] }
        wouldFireCount: { type: integer }
        updatedBy: { type: ["string", "null"] }
        updatedAt: { type: string, format: date-time }
    AlertSummary:
      type: object
      required: [total, fired, muted, suppressed, clear, quantityAtStake]
      properties:
        total: { type: integer }
        fired: { type: integer }
        muted: { type: integer }
        suppressed: { type: integer }
        clear: { type: integer }
        quantityAtStake: { type: number }
    AlertsResponse:
      type: object
      required: [month, segment, summary, alerts, thresholds]
      properties:
        month: { type: string }
        segment: { type: string, enum: [PTMT, Plumbing] }
        summary: { $ref: "#/components/schemas/AlertSummary" }
        alerts:
          type: array
          items: { $ref: "#/components/schemas/Alert" }
        thresholds:
          type: array
          items: { $ref: "#/components/schemas/AlertThreshold" }
    AlertHistoryEvent:
      type: object
      required: [id, alertId, code, segment, month, state, value, threshold, difference, quantity, message, details, sourceLinks, occurredAt, actor, action]
      properties:
        id: { type: integer }
        alertId: { type: integer }
        code: { type: string }
        segment: { type: string }
        month: { type: string }
        state: { type: string }
        value: { type: ["number", "null"] }
        threshold: { type: ["number", "null"] }
        difference: { type: ["number", "null"] }
        quantity: { type: number }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        sourceLinks:
          type: array
          items: { $ref: "#/components/schemas/AlertSourceLink" }
        occurredAt: { type: string, format: date-time }
        actor: { type: ["string", "null"] }
        action: { type: string }
    AlertHistoryResponse:
      type: object
      required: [segment, history]
      properties:
        segment: { type: string, enum: [PTMT, Plumbing] }
        history:
          type: array
          items: { $ref: "#/components/schemas/AlertHistoryEvent" }
    MuteAlertRequest:
      type: object
      required: [reason, mutedUntil]
      properties:
        reason: { type: string, minLength: 1 }
        mutedUntil: { type: string, format: date-time }
    UpdateAlertThresholdRequest:
      type: object
      required: [segment, value, reason]
      properties:
        segment: { type: string, enum: [PTMT, Plumbing] }
        value: { type: number, minimum: 0 }
        reason: { type: string, minLength: 1 }
    AlertActionResponse:
      type: object
      required: [ok, alert]
      properties:
        ok: { type: boolean }
        alert: { $ref: "#/components/schemas/Alert" }
    AlertThresholdActionResponse:
      type: object
      required: [ok, threshold]
      properties:
        ok: { type: boolean }
        threshold: { $ref: "#/components/schemas/AlertThreshold" }